# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/open-apis/core/rbac/tenant/permission": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "open-apis/core"
                ],
                "summary": "获取权限项列表方法(TenantAccessToken)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "操作类型",
                        "name": "operation_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "权限名称",
                        "name": "permission_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "资源类型",
                        "name": "resource_type",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}